refactor(harness): curate grounding diagnostics, skills catalog and goal chrome into spi; shell ledger 18→15
