+++ tools/interactive_shell/actions/slash.py
@@ -7,8 +7,8 @@ from typing import Any
 
 from rich.markup import escape
 
-from core.agent_harness.spi import session_terminal, set_auto_command
 from core.agent_harness.session.terminal_access import exclusive_stdin_active
+from core.agent_harness.spi import session_terminal, set_auto_command
 from core.agent_harness.tools.tool_context import (
     ActionToolContext,
     capability_available_from_sources,
